+++ packages/pi-session-tools/CHANGELOG.md
@@ -1,5 +1,28 @@
 # Changelog
 
+## [0.4.0](https://github.com/maplezzk/pi-extensions/compare/pi-session-tools-v0.3.1...pi-session-tools-v0.4.0) (2026-08-19)
+
+
+### Features
+
+* **session-tools:** streamline and enforce session squash ([dea7e72](https://github.com/maplezzk/pi-extensions/commit/dea7e727a1bd2ea2d8a74dbdaf43dcfdc7c68639))
+* **session-tools:** 将摘要直接传入 session_squash ([080d777](https://github.com/maplezzk/pi-extensions/commit/080d7773ef83b44de83eeeccde3d6c1958f14e93))
+* **session-tools:** 支持强制会话压缩 ([55e8c49](https://github.com/maplezzk/pi-extensions/commit/55e8c493ccb28d6a2b3209aa41c99c10f560e255))
+
+
+### Bug Fixes
+
+* **session-tools:** 强制阈值改用数字比例 ([ac22144](https://github.com/maplezzk/pi-extensions/commit/ac221446f8e04f7711b513eae3b30573d282b54c))
+* **session-tools:** 支持从更早索引重新压缩 ([ac66fc0](https://github.com/maplezzk/pi-extensions/commit/ac66fc0242acdf3e320b3bdf683296056d829647))
+
+## [0.3.1](https://github.com/maplezzk/pi-extensions/compare/pi-session-tools-v0.3.0...pi-session-tools-v0.3.1) (2026-08-18)
+
+
+### Bug Fixes
+
+* suppress reminders after aborted or failed runs ([a23d950](https://github.com/maplezzk/pi-extensions/commit/a23d950c716909a5e563f19d616e150a529fd54e))
+* 忽略异常终止的自动提醒 ([743ac5a](https://github.com/maplezzk/pi-extensions/commit/743ac5a7886165fcc4f98150f648e70202b4f4fe))
+
 ## [0.3.0](https://github.com/maplezzk/pi-extensions/compare/pi-session-tools-v0.2.0...pi-session-tools-v0.3.0) (2026-08-14)